requirements.txt
# ChipMASRAG 依赖包

# 核心依赖
numpy>=1.21.0
scipy>=1.7.0
torch>=1.10.0
torch-geometric>=2.0.0

# RAG和嵌入
sentence-transformers>=2.2.0
transformers>=4.20.0

# 数据处理
pandas>=1.3.0
networkx>=2.6.0

# 配置文件
pyyaml>=5.4.0

# OpenRoad接口（如果使用Python API）
# 注意：OpenRoad可能需要单独安装

# 可视化
matplotlib>=3.4.0
seaborn>=0.11.0

# 实验和日志
tensorboard>=2.7.0
tqdm>=4.62.0

# 工具
tqdm>=4.62.0
psutil>=5.8.0  # 资源监控

# 可选：强化学习框架（如果需要）
# stable-baselines3>=1.5.0  # 如果需要使用stable-baselines3

# 开发和测试
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0

